debugging

+++ 0x00-python_variable_annotations/6-sum_mixed_list.py
@@ -1,11 +1,10 @@
 #!/usr/bin/env python3
-from typing import List, Union
-
 """
   type annotation basics
 """
 
+from typing import List, Union
 
-def sum_mixed_list(mxd_lst: List[Union[float, int]]) -> float:
+def sum_mixed_list(mxd_lst: List[Union[int, float]]) -> float:
     """returns the sum of mxd list"""
     return sum(mxd_lst)
